split GVRResolver out of dynamic helper
