+++ pkg/util/dynamichelper/dynamichelper.go
@@ -5,9 +5,7 @@ package dynamichelper
 
 import (
 	"context"
-	"net/http"
 	"reflect"
-	"strings"
 
 	"github.com/sirupsen/logrus"
 	"github.com/ugorji/go/codec"
@@ -15,18 +13,16 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/client-go/discovery"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/util/retry"
 
-	"github.com/Azure/ARO-RP/pkg/api"
 	"github.com/Azure/ARO-RP/pkg/util/cmp"
-	kadiscovery "github.com/Azure/ARO-RP/pkg/util/dynamichelper/discovery"
 )
 
 type Interface interface {
-	RefreshAPIResources() error
+	GVRResolver
+
 	CreateOrUpdate(ctx context.Context, obj *unstructured.Unstructured) error
 	Delete(ctx context.Context, groupKind, namespace, name string) error
 	Ensure(ctx context.Context, objs ...*unstructured.Unstructured) error
@@ -35,11 +31,12 @@ type Interface interface {
 }
 
 type dynamicHelper struct {
+	GVRResolver
+
 	log *logrus.Entry
 
-	restconfig   *rest.Config
-	dyn          dynamic.Interface
-	apiresources []*metav1.APIResourceList
+	restconfig *rest.Config
+	dyn        dynamic.Interface
 }
 
 func New(log *logrus.Entry, restconfig *rest.Config) (Interface, error) {
@@ -49,7 +46,13 @@ func New(log *logrus.Entry, restconfig *rest.Config) (Interface, error) {
 	}
 
 	var err error
-	dh.dyn, err = dynamic.NewForConfig(dh.restconfig)
+
+	dh.GVRResolver, err = NewGVRResolver(log, restconfig)
+	if err != nil {
+		return nil, err
+	}
+
+	dh.dyn, err = dynamic.NewForConfig(restconfig)
 	if err != nil {
 		return nil, err
 	}
@@ -57,96 +60,11 @@ func New(log *logrus.Entry, restconfig *rest.Config) (Interface, error) {
 	return dh, nil
 }
 
-func (dh *dynamicHelper) RefreshAPIResources() error {
-	var cli discovery.DiscoveryInterface
-	cli, err := discovery.NewDiscoveryClientForConfig(dh.restconfig)
-	if err != nil {
-		return err
-	}
-	cli = kadiscovery.NewCacheFallbackDiscoveryClient(dh.log, cli)
-
-	_, dh.apiresources, err = cli.ServerGroupsAndResources()
-	if discovery.IsGroupDiscoveryFailedError(err) {
-		// Some group discovery failed; dh.apiresources will have all the ones
-		// that worked. This error can happen with a misconfigured apiservice,
-		// for example. Log it and try to keep going.
-		dh.log.Warn(err)
-		return nil
-	}
-	return err
-}
-
-func (dh *dynamicHelper) findGVR(groupKind, optionalVersion string) (*schema.GroupVersionResource, error) {
-	if dh.apiresources == nil {
-		err := dh.RefreshAPIResources()
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	var matches []*schema.GroupVersionResource
-	for _, apiresources := range dh.apiresources {
-		gv, err := schema.ParseGroupVersion(apiresources.GroupVersion)
-		if err != nil {
-			// this returns a fmt.Errorf which will result in a 500
-			// in this case, this seems correct as the GV in kubernetes is wrong
-			return nil, err
-		}
-		if optionalVersion != "" && gv.Version != optionalVersion {
-			continue
-		}
-		for _, apiresource := range apiresources.APIResources {
-			if strings.ContainsRune(apiresource.Name, '/') { // no subresources
-				continue
-			}
-
-			gk := schema.GroupKind{
-				Group: gv.Group,
-				Kind:  apiresource.Kind,
-			}
-
-			if strings.EqualFold(gk.String(), groupKind) {
-				return &schema.GroupVersionResource{
-					Group:    gv.Group,
-					Version:  gv.Version,
-					Resource: apiresource.Name,
-				}, nil
-			}
-
-			if strings.EqualFold(apiresource.Kind, groupKind) {
-				matches = append(matches, &schema.GroupVersionResource{
-					Group:    gv.Group,
-					Version:  gv.Version,
-					Resource: apiresource.Name,
-				})
-			}
-		}
-	}
-
-	if len(matches) == 0 {
-		return nil, api.NewCloudError(
-			http.StatusBadRequest, api.CloudErrorCodeNotFound,
-			"", "The groupKind '%s' was not found.", groupKind)
-	}
-
-	if len(matches) > 1 {
-		var matchesGK []string
-		for _, match := range matches {
-			matchesGK = append(matchesGK, groupKind+"."+match.Group)
-		}
-		return nil, api.NewCloudError(
-			http.StatusBadRequest, api.CloudErrorCodeInvalidParameter,
-			"", "The groupKind '%s' matched multiple groupKinds (%s).", groupKind, strings.Join(matchesGK, ", "))
-	}
-
-	return matches[0], nil
-}
-
 // CreateOrUpdate does nothing more than an Update call (and a Create if that
 // call returned 404).  We don't add any fancy behaviour because this is called
 // from the Geneva Admin context and we don't want to get in the SRE's way.
 func (dh *dynamicHelper) CreateOrUpdate(ctx context.Context, o *unstructured.Unstructured) error {
-	gvr, err := dh.findGVR(o.GroupVersionKind().GroupKind().String(), o.GroupVersionKind().Version)
+	gvr, err := dh.Resolve(o.GroupVersionKind().GroupKind().String(), o.GroupVersionKind().Version)
 	if err != nil {
 		return err
 	}
@@ -161,7 +79,7 @@ func (dh *dynamicHelper) CreateOrUpdate(ctx context.Context, o *unstructured.Uns
 }
 
 func (dh *dynamicHelper) Delete(ctx context.Context, groupKind, namespace, name string) error {
-	gvr, err := dh.findGVR(groupKind, "")
+	gvr, err := dh.Resolve(groupKind, "")
 	if err != nil {
 		return err
 	}
@@ -185,7 +103,7 @@ func (dh *dynamicHelper) Ensure(ctx context.Context, objs ...*unstructured.Unstr
 }
 
 func (dh *dynamicHelper) ensureOne(ctx context.Context, o *unstructured.Unstructured) error {
-	gvr, err := dh.findGVR(o.GroupVersionKind().GroupKind().String(), o.GroupVersionKind().Version)
+	gvr, err := dh.Resolve(o.GroupVersionKind().GroupKind().String(), o.GroupVersionKind().Version)
 	if err != nil {
 		return err
 	}
@@ -214,7 +132,7 @@ func (dh *dynamicHelper) ensureOne(ctx context.Context, o *unstructured.Unstruct
 }
 
 func (dh *dynamicHelper) Get(ctx context.Context, groupKind, namespace, name string) (*unstructured.Unstructured, error) {
-	gvr, err := dh.findGVR(groupKind, "")
+	gvr, err := dh.Resolve(groupKind, "")
 	if err != nil {
 		return nil, err
 	}
@@ -223,7 +141,7 @@ func (dh *dynamicHelper) Get(ctx context.Context, groupKind, namespace, name str
 }
 
 func (dh *dynamicHelper) List(ctx context.Context, groupKind, namespace string) (*unstructured.UnstructuredList, error) {
-	gvr, err := dh.findGVR(groupKind, "")
+	gvr, err := dh.Resolve(groupKind, "")
 	if err != nil {
 		return nil, err
 	}
